sdcard: if distro_blank is not present use normal

+++ tools/setup_sdcard.sh
@@ -217,9 +217,17 @@ distro_bootloader () {
 
 	if [ "x${oem_blank_eeprom}" = "xenable" ] ; then
 		if [ "${conf_bl_distro_blank_SPL}" ] ; then
-			cp -v ./${conf_bl_distro_blank_SPL} ${TEMPDIR}/dl/oem/
-			blank_SPL=${spl_name}
-			echo "blank_SPL Bootloader: ${conf_bl_distro_blank_SPL}"
+			if [ -f ${conf_bl_distro_blank_SPL} ] ; then
+				cp -v ./${conf_bl_distro_blank_SPL} ${TEMPDIR}/dl/oem/
+				blank_SPL=${spl_name}
+				echo "blank_SPL Bootloader: ${conf_bl_distro_blank_SPL}"
+			else
+				if [ "${conf_bl_distro_SPL}" ] ; then
+					cp -v ./${conf_bl_distro_SPL} ${TEMPDIR}/dl/oem/
+					blank_SPL=${spl_name}
+					echo "SPL Bootloader: ${conf_bl_distro_SPL}"
+				fi
+			fi
 		else
 			if [ "${conf_bl_distro_SPL}" ] ; then
 				cp -v ./${conf_bl_distro_SPL} ${TEMPDIR}/dl/oem/
@@ -229,9 +237,17 @@ distro_bootloader () {
 		fi
 
 		if [ "${conf_bl_distro_blank_UBOOT}" ] ; then
-			cp -v ./${conf_bl_distro_blank_UBOOT} ${TEMPDIR}/dl/oem/
-			blank_UBOOT=${boot_name}
-			echo "blank_UBOOT Bootloader: ${conf_bl_distro_blank_UBOOT}"
+			if [ -f ${conf_bl_distro_blank_UBOOT} ] ; then
+				cp -v ./${conf_bl_distro_blank_UBOOT} ${TEMPDIR}/dl/oem/
+				blank_UBOOT=${boot_name}
+				echo "blank_UBOOT Bootloader: ${conf_bl_distro_blank_UBOOT}"
+			else
+				if [ "${conf_bl_distro_UBOOT}" ] ; then
+					cp -v ./${conf_bl_distro_UBOOT} ${TEMPDIR}/dl/oem/
+					blank_UBOOT=${boot_name}
+					echo "UBOOT Bootloader: ${conf_bl_distro_UBOOT}"
+				fi
+			fi
 		else
 			if [ "${conf_bl_distro_UBOOT}" ] ; then
 				cp -v ./${conf_bl_distro_UBOOT} ${TEMPDIR}/dl/oem/
